Decrypt data on loading servers instead of waiting for open the connection
This makes easier. Because we don't care about decrypt them later. So is less code to change.
Also provides an easier way to export the decrypted data.
